Melhorias de segurança e integridade dos dados de usuário

+++ src/main/java/br/com/projeto/contratados/domain/service/usuario/ExperienciaService.java
@@ -2,6 +2,7 @@ package br.com.projeto.contratados.domain.service.usuario;
 
 import br.com.projeto.contratados.config.exception.excecoes.ExperienciaNaoEncontradaException;
 import br.com.projeto.contratados.config.exception.excecoes.UsuarioNaoEncontradoException;
+import br.com.projeto.contratados.config.security.TokenService;
 import br.com.projeto.contratados.domain.entity.usuario.Experiencia;
 import br.com.projeto.contratados.domain.entity.usuario.Usuario;
 import br.com.projeto.contratados.domain.repository.usuario.ExperienciaRepository;
@@ -23,13 +24,21 @@ public class ExperienciaService {
     @Autowired
     private UsuarioRepository usuarioRepository;
 
+    @Autowired
+    private TokenService tokenService;
+
+    private Integer getIdUsuario(){
+
+        return tokenService.getAuthenticatedUsuario();
+    }
+
     public Experiencia cadastrar(ExperienciaRequest form) {
 
-        var experiencia = form.converte();
-
-        Optional<Usuario> optional = usuarioRepository.findById(experiencia.getUsuario().getId());
+        Optional<Usuario> optional = usuarioRepository.findById(getIdUsuario());
         if (optional.isEmpty())
             throw new UsuarioNaoEncontradoException("Usuário não encontrado, não foi possível adicinar uma formação");
+
+        var experiencia = form.converte(optional.get());
 
         return experienciaRepository.save(experiencia);
     }
@@ -37,9 +46,9 @@ public class ExperienciaService {
     public Page<Experiencia> listar(String descricao, Pageable paginacao) {
 
         if (descricao == null)
-            return experienciaRepository.findAll(paginacao);
+            return experienciaRepository.findByUsuarioId(getIdUsuario(), paginacao);
 
-        return experienciaRepository.findByDescricao(descricao, paginacao);
+        return experienciaRepository.findByDescricaoAndUsuarioId(descricao, getIdUsuario(), paginacao);
     }
 
     public Experiencia atualizar(Integer id, AtualizacaoExperienciaRequest form) {
@@ -47,6 +56,9 @@ public class ExperienciaService {
         Optional<Experiencia> optional = experienciaRepository.findById(id);
         if (optional.isEmpty())
             throw new ExperienciaNaoEncontradaException("Experiência não encontrada, não pode ser atualizada");
+
+        if (!optional.get().getUsuario().getId().equals(getIdUsuario()))
+            throw new UsuarioNaoEncontradoException("Usuário não encontrado, a experiência não pode ser atualizada");
 
         var experiencia = form.atualizacaoExperienciaForm(id, experienciaRepository);
         return experienciaRepository.save(experiencia);
@@ -56,6 +68,9 @@ public class ExperienciaService {
         Optional<Experiencia> optional = experienciaRepository.findById(id);
         if(optional.isEmpty())
             throw new ExperienciaNaoEncontradaException("Experiência não encontrada, não pode ser deletada");
+
+        if (!optional.get().getUsuario().getId().equals(getIdUsuario()))
+            throw new UsuarioNaoEncontradoException("Usuário não encontrado, a experiência não pode ser atualizada");
 
         var experiencia = experienciaRepository.getOne(id);
         experienciaRepository.deleteById(id);
